# docs/swagger.yaml
definitions:
  model.Detalle:
    properties:
      descripcion:
        type: string
      estado:
        type: string
      idAuxiliar:
        type: string
      idDetalle:
        type: integer
      idMantenimiento:
        type: string
      nombre:
        type: string
      usuarioRegistro:
        type: string
    type: object
  model.Empleado:
    properties:
      apellidos:
        type: string
      celular:
        type: string
      clave:
        type: string
      count:
        type: integer
      detalle:
        $ref: '#/definitions/model.Detalle'
      direccion:
        type: string
      email:
        type: string
      estado:
        type: integer
      fechaNacimiento:
        type: string
      huella:
        type: string
      id:
        type: integer
      idEmpleado:
        type: string
      idRol:
        type: integer
      nombres:
        type: string
      numeroDocumento:
        type: string
      puesto:
        type: integer
      rol:
        $ref: '#/definitions/model.Rol'
      sexo:
        type: integer
      sistema:
        type: boolean
      telefono:
        type: string
      tipoDocumento:
        type: integer
      usuario:
        type: string
    type: object
  model.Empresa:
    properties:
      celular:
        type: string
      certificadoClave:
        type: string
      certificadoRuta:
        type: string
      claveApiSunat:
        type: string
      claveSol:
        type: string
      condiciones:
        type: string
      domicilio:
        type: string
      email:
        type: string
      giroComercial:
        type: integer
      idApiSunat:
        type: string
      idEmpresa:
        type: integer
      image:
        items:
          type: integer
        type: array
      imagenRuta:
        type: string
      nombre:
        type: string
      nombreComercial:
        type: string
      numeroDocumento:
        type: string
      paginaWeb:
        type: string
      razonSocial:
        type: string
      telefono:
        type: string
      terminos:
        type: string
      tipoDocumento:
        type: integer
      ubigeo:
        type: integer
      usuarioSol:
        type: string
    type: object
  model.Error:
    properties:
      message:
        type: string
    type: object
  model.Rol:
    properties:
      id:
        type: integer
      idRol:
        type: integer
      nombre:
        type: string
      sistema:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /empleado:
    get:
      consumes:
      - application/json
      description: Ruta usada para traer datos relevante al momento de realizar una
        edición
      parameters:
      - description: Id del empleado
        in: query
        name: idEmpleado
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Empleado'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      summary: Obtener empleado po su Id
      tags:
      - Empleado
  /empleados:
    get:
      consumes:
      - application/json
      description: Listado de empleados o usuario con los datos principales
      parameters:
      - description: Opciones de filtro 0-libre 1-para iniciar la busqueda
        in: query
        name: opcion
        required: true
        type: integer
      - description: Datos para el filtro
        in: query
        name: search
        type: string
      - description: Inicio de la paginación
        in: query
        name: posicionPagina
        required: true
        type: integer
      - description: Filas por paginación
        in: query
        name: filasPorPagina
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Empleado'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      summary: Lista de empleados o usarios del sistema
      tags:
      - Empleado
  /empresa:
    get:
      consumes:
      - application/json
      description: Obtener la información de la empresa
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Empresa'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      summary: Para obtener la empresa actual
      tags:
      - Empresa
  /lista-detalle-idmantenimiento:
    get:
      consumes:
      - application/json
      description: Obtener el listado de detalle por id mantenimiento
      parameters:
      - description: Opciones de filtro 0-libre 1-para excluir el nombre al inciar
          la busqueda
        in: query
        name: opcion
        required: true
        type: string
      - description: Id del matenimiento
        in: query
        name: idMantenimiento
        required: true
        type: string
      - description: nombre a excluir
        in: query
        name: nombre
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Detalle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      summary: Obtener detalle
      tags:
      - Detalle
  /listar-roles:
    get:
      consumes:
      - application/json
      description: Obtener el listado de roles para los modulos que necesiten
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Rol'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      summary: Lista de roles
      tags:
      - Rol
  /login:
    get:
      consumes:
      - application/json
      description: Iniciar Sesión del Empleado
      parameters:
      - description: Usuario para iniciar sesión
        in: query
        name: usuario
        required: true
        type: string
      - description: Clave para iniciar sesión
        in: query
        name: clave
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Empleado'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Error'
      summary: Para el inicio de sesión
      tags:
      - Empleado
swagger: "2.0"